+++ components/layout/construct.rs
@@ -51,7 +51,7 @@ use style::computed_values::position;
 use style::context::SharedStyleContext;
 use style::logical_geometry::Direction;
 use style::properties::{self, ServoComputedValues};
-use style::selector_impl::{PseudoElement, RestyleDamage};
+use style::selector_parser::{PseudoElement, RestyleDamage};
 use style::servo::restyle_damage::{BUBBLE_ISIZES, RECONSTRUCT_FLOW};
 use style::stylist::Stylist;
 use table::TableFlow;
