+++ components/script/dom/element.rs
@@ -92,7 +92,7 @@ use style::properties::{DeclaredValue, Importance};
 use style::properties::{PropertyDeclaration, PropertyDeclarationBlock, parse_style_attribute};
 use style::properties::longhands::{background_image, border_spacing, font_family, font_size, overflow_x};
 use style::restyle_hints::RESTYLE_SELF;
-use style::selector_impl::{NonTSPseudoClass, RestyleDamage, ServoSelectorImpl};
+use style::selector_parser::{NonTSPseudoClass, RestyleDamage, ServoSelectorImpl};
 use style::sink::Push;
 use style::stylist::ApplicableDeclarationBlock;
 use style::values::CSSFloat;
