+++ components/style/lib.rs
@@ -114,7 +114,7 @@ pub mod parser;
 pub mod refcell;
 pub mod restyle_hints;
 pub mod rule_tree;
-pub mod selector_impl;
+pub mod selector_parser;
 pub mod stylist;
 #[cfg(feature = "servo")] #[allow(unsafe_code)] pub mod servo;
 pub mod sequential;
