+++ components/style/servo/selector_parser.rs
@@ -7,8 +7,8 @@ use attr::{AttrIdentifier, AttrValue};
 use cssparser::ToCss;
 use element_state::ElementState;
 use restyle_hints::ElementSnapshot;
-use selector_impl::{ElementExt, PseudoElementCascadeType, TheSelectorImpl};
-use selector_impl::{attr_equals_selector_is_shareable, attr_exists_selector_is_shareable};
+use selector_parser::{ElementExt, PseudoElementCascadeType, TheSelectorImpl};
+use selector_parser::{attr_equals_selector_is_shareable, attr_exists_selector_is_shareable};
 use selectors::{Element, MatchAttrGeneric};
 use selectors::parser::{AttrSelector, ParserContext, SelectorImpl};
 use std::fmt;
